refactor(entity-decode): address review — memoise compiled regex, dedupe client decoder
- cache the compiled pattern keyed on the config-derived string, so per-item
  seeder calls stop re-compiling identical regexes
- move the client-side single-pass decoder to src/utils/decode-entities.ts
  (utils <- components per AGENTS.md) and delegate from both
  CountryDeepDivePanel call sites
