+++ src/components/CountryDeepDivePanel-news-utils.ts
@@ -1,18 +1,7 @@
 import type { NewsItem } from '../types';
+import { decodeHtmlEntities } from '../utils/decode-entities';
 
-export function decodeHtmlEntities(text: string): string {
-  // Single-pass decode (issue #5436): decode exactly one entity level.
-  return text.replace(/&(?:amp|lt|gt|quot|#(\d+)|#[xX]([0-9a-fA-F]+));/g, (m, dec, hex) => {
-    if (dec !== undefined || hex !== undefined) {
-      const cp = hex !== undefined ? parseInt(hex, 16) : Number(dec);
-      if (cp === 39) return "'";
-      if (cp === 47) return '/';
-      return m;
-    }
-    const named: Record<string, string> = { '&amp;': '&', '&lt;': '<', '&gt;': '>', '&quot;': '"' };
-    return named[m] ?? m;
-  });
-}
+export { decodeHtmlEntities };
 
 export function normalizeHeadlineKey(title: string): string {
   return decodeHtmlEntities(title)
